requirements.txt
python-dotenv
Flask
Flask-Cors
pytest
pandas
mlflow
notebook
numpy
scikit-learn
matplotlib
python-box
pyYAML
types-PyYAML
tqdm
ensure
joblib
pymongo[srv]==3.12
pymongo
certifi
dill

# -e .